Fixed handling of zeroed inode tables.
A zeroed inode table does not necessarily imply that it has no valid
inodes, only that the unused inodes are zero. The check for the zeroed
flag should therefore not determine whether or not a block group's inode
table is parsed. This should be determined by the bg_free_inodes_count
(skip if it is equal to the superblock's s_inodes_per_group).

This change fixes that.