Added free-extent updating on file extent moving.
When moving multiple file extents, it might be useful if the tree of
free extents actually represents the free spaces on disk. This change
introduces the allocate_space() and deallocate_space() functions that
update the trees. It also updates the move_file_extent() function to use
these new functions.

Note that I've had to introduce a new free-extent tree ordered by
starting block to allow for merging of adjacent extents.