Fixed potential bug in whole-file moving.
Because of the merging of extents, the extent numbers may change after moving
an extent. When moving several extents of a file, this could cause various
extents to be moved to the incorrect locations, leading to a non-defragmented
filesystem or a program crash. I don't think this could have caused data loss.

In any event, here is a fix whereby the extent moving function returns the
number of extent entries by which the moved extent has moved in the inode
extent list.