Fixed order of free blocks in extent tree.
The old tree had the free extents in descending order, which will cause
problems when we want to move a file that spans multiple extents (as we want
to get to the next free extent, but we might get the previous one).
This puts the extents in proper ascending order.