Added method for moving files that span multiple free extents.
This change adds support for splitting extents that do not fit within
the specified extent, and automatically moving the remainder to the
next free extent on disk. Behaviour when moving past the last free extent
on disk is undefined.
This also fixes calculations for indirect block writing. The errors only
show up when you start writing in the middle of an indirect block which
never happened before this.