Fixed data corruption on errors.
The system is supposed to only write pointers to data that has been
correctly written, but the indirect blocks would have a portion of them
zero-filled when an extent does not cover the entire block. Normally
this does not cause problems because the next extent is soon after moved
and written back (updating the metadata blocks), but in failure situations
this does not happen and data is lost.

This fixes that situation by not touching the parts of indirect metadata
blocks that do not refer to the extent being written.