Added support for reading ext4 extents.
Structure: metadata included as normal extents at the beginning of the
inode extent list. The inode now has a field called num_extent_index_blocks.
If this field is nonzero, it indicated how many logical blocks at the
beginning of the inode are extent indexes. If this field is zero, either
the inode has no blocks or uses direct addressing.

Writeback is not yet supported, and may be a little hairy because the
physical position of a block reference cannot be computed from the
logical block number, and also because the extent tree structure may
change. It might be easier to just rewrite the entire index every time.