Separated extent metadata from file data.
When defragmenting a file, the extent tree might shrink (because we need
fewer extents, so fewer nodes, so we might fit things into the inode
rather than needing an index block. Because of this, the position we
so carefully chose to align with another file might change when we don't
need the extent tree at the start or and end of the file.
Separating the extent tree is the only meaningful way to resolve this I
could come up with.