Added support for writing back extent metadata
It current allocates (very stupidly, just picks the first n blocks out of
the biggest free space on disk) new space for the new metadata, and writes
the indexes and extents to that, then points the inode to the new data and
deallocates the old metadata. This seems to be the safest way (only change of
data corruption is if the system fails during sync of the inode block, or
if the kernel decides to sync halfway into the write).

This method, hoewever, requires some additional space and may fragment the
filesystem if it's very low on space. It also prevents intentional
defragmenting of the metadata blocks for whatever reason.

To facilitate this system, there are two new functions for freespace
management: allocate_blocks and deallocate_blocks to allocate a specific
number of blocks in as few extents as possible, returned in a struct
allocation. deallocate_blocks deallocates a given struct allocation.