Added display of potentially fragmented inodes before asking inode number
Just dumping the whole tree displays a lot of information about files
that are just not fragmented, which is completely pointless and obscures
the files that are fragmented.

This adds a listing of the fragmented inodes, along with the number
of extents and the number of blocks in the inode. It displays both
the actual data and the metadata allocations (if applicable).