Added basic algorithm for automatically defragmenting an inode.
This one is really simple, it just picks the smallest number n of
the largest fragments, so that the sum of the fragments is enough
to fit the file. If this n is smaller than the current number of
extents in the file, it defragments to that combination. Otherwise,
it does nothing to the file.