Fixed use after free and minor (unexploitable) buffer overrun
The use after free was a simple case of returning a value from a
structure that has already been freed. Fixed that by reading the
returned value before freeing.

The buffer overrun was part of the extent walking code in metadata
writeback, where it would seek past the last extent (because the next
block that would be written would be in the next extent, even though
no more blocks will be written) and preread the starting blocks number.

Because this value is never used (as no more blocks will be written), and
only reads are performed, this would never leak information or cause
any data corruption (and is therefore unexploitable).

Fixed it anyway because it's poor style and messes up the valgrind log.