Let allocate_blocks allocate more than one extent.
It's still rather naive, because it always allocates the biggest extents
it can find, but it's better than it was. Note that this is not very
well tested and will probably reveal bugs in other parts of the program
that worked correctly when only one extent could be returned. Beware
of possible data corruption from this.