Fixed data corrupting bug in multi-extent moving
We were splitting the extents at the wrong place (forgot to subtract
a 1), which caused the program to overwrite the block immediatly
following a newly written extent in some cases (specifically, when
the original extent had to be split because it didn't fit). This
caused data corruption if that block was already in use.