Added checksum updating of group descriptors
Fixes the need to run e2fsck after every defrag before mounting. Also,
if any data was moved into a block group with an uninitialized inode table,
there was also the problem that e2fsck treated it as initialized because
it can't trust the descriptor with the bad checksum, which meant a bunch
of invalid inodes and other such problems.