Added updating of the inodes i_blocks field on metadata changes
When, after defragmenting, there are fewer extents than before,
sometimes the extent tree can take fewer blocks than it was before. When
this happens, the i_blocks field of the inode needs to be updated
or e2fsck complains (I don't think it will cause any problems in the
kernel as the extra blocks will just be ignored because of the i_size,
but better safe than sorry).

This should also work when more blocks are needed, but that never
happens in this program.

Implementation note: the i_blocks update is not synced because
it doesn't cause any ordering problems if it doesn't happen.