Added basic free space consolidation.
Rather basic, and doesn't work that well but it's a start.
Note that the algorithm to find a proper free space is rather
inefficient (performing O(n log n), rather than O(log n)).

This does not break the fundamental terminating requirement of the
algorithms used (i.e. no matter what actions you take, eventually
you end up in a state where none of the algorithms can proceed) and
I'd like to keep it that way.