Hacked fix for bug when moving metadata
Moving metadata extents that lie inbetween two free extents (so that the
extents are merged after the move) caused a use-after-free that crashed
the program and on at least one occasion caused data loss. Fixed that by
not moving metadata (but rather calling the write_metadata function when
we want to move it). A real fix is still in the making.