Fixed metadata writeback after move
When the metadata extent moved was in an index pointed to by the last
entry in a higher index, the program did not update the on-disk structure
of the metadata to point to the new index. This could easily cause
data corruption if the old index block was later overwritten (but still
pointed to).