Added display of number of biggest extents to interactive mode
On disks that do not use the flex_bg feature, there are typically
a lot of free extents around 32765 (depending on the inode block count),
representing the biggest free extents on the disk. Since only displaying
the size of this extent is not that useful for determining whether a file
can be maximally defragmented, I added the count of these extents as well.