Fixed two data-corrupting bugs
First, when updating a single extent in an inode with indirect-block
addressing, the remaining blocks would be zereoed out (which used to
be fine when the next extent would be written immediatly after, but
causes data loss when compacting free space and not moving the next
extents).

Then, another bug surfaced corrupting files with sparse extents. The
is_sparse function got the greater than/smaller than exactly the wrong
way around. Fixed that as well so I could properly test the other bug.

Haven't tested these individually, so I figured it'd be better
commiting then in one go.