Added confirmation before inode move in interactive mode
The program now tells the user how many fragments the new inode
would have, and lets the user decide whether this is good enough.
This is particularly useful on filesystems without flex_bg where
complete defragmentation of large files is impossible, and the
user can decide how many fragments for a file is acceptable.