Added noninteractive whole-disk mode.
This is now the default for the program. The old interactive mode can still
be used by passing the -i option.
The algorithm used by this non-interactive mode is still very stupid. It
just tries to optimize every inode, and once it's finished, if anything
is still suboptimal, it tries to pack the free space and try again. A very
stupid part of this is that 'optimal' is defined as having only one extent,
which means that disks with large files without large flex_bg groups
will have their free space fully consolidated.