GPL'ed everything.
I'm releasing this program under version 2 of the GPL. I won't do
version 3 or 'any later version', partly because I don't completely
trust the FSF to make the right decisions, and partly because the crc16
library was copied from the linux kernel and can only be released under
GPLv2.

I hope the red-black-tree implementation won't be a problem, because it
is released under any later version (and I didn't write it, so I can't
change it). I'm no lawyer, but I think the 'you can't impose any further
restrictions' part of the license might be a problem. If it is, I can
probably find another implementation somewhere.

I'm also not quite sure what to do with the binary images distributed as
part of the testcases because they don't have any source code. Oh well,
that's for anyone else who wants to distribute them to work out.