Added testcase for in place defragmenting.
The program should be able to defragmentn files leaving some existing
extents in their original location. The incremental improvement feature
is a good step in this direction, but I don't think it's quite general
enough. This testcase tests a situation in which the only possible
defragmentation is obtained by leaving the first extent in place, and
moving the second (two-block) extent one block to the left. While this
is a rather extreme situation, I think the program should be able to
deal with this (especially on very full drives).