Fixed data-corrupting bug in free space consolidation
When moving an extent to consolidate free space, the uninit field in the
new extent was not being properly initialized. This means that, if the
memory in which the data about the new extent was previously freed from
an uninitialized extent, the extent being moved would become
uninitialized itself. This would cause that extent to appear zeroed if
later read. Fixed that by copying the uninit field from the old extent.