Fixed bug causing non-terminating testcases
This was caused by two problems:

First: after defragmenting an inode in do_whole_disk, the check whether
the resulting placement is still fragmented did not consider the
possibility of an uninitialized extent and an initialized one being
consecutive, and wrongly flagged these as still fragmented.
Second: because the new placement was considered non-optimal, it tried
to consolidate free space. This algorithm also did not consider the
split-but-consecutive case and moved one of the extents away, so that
next time in the loop the inode would have to be defragmented again,
etc...

Two fixed applied: The first problem was corrected by fixing
is_fragmented (and using it in all the right places, we were still
checking raw extent counts in some places). The second problem is
patched up by refusing to move an extent if this would separate it from
a preceding extent of the same inode. Some later time we should add code
to move the two (or more) extents together.