Fixed program lockup when trying to improve some files.
A file that cannot be improved further by the improvement algorithm
locks up the program (as demonstrated by test t1111-full). Fixed that
returning the original allocation when no changed were made, rather than
an exact copy. This way the caller can easily tell that the inode cannot
be improved and it should stop trying.