Made tests less locale-specific
I was rather surprised when all of the tests suddenly failed on my
new laptop, only to discover that it was because the locale was set to
dutch, so the phrase "0.0% non-contiguous" does not occur in the
e2fsck output. I think just setting LC_ALL to C before running a test
or infra command fixes it (at least it does for me).