Reject unclean filesystems, and mark active filesystems unclean.
Defragmenting a filesystem that was not cleanly umounted (or has errors)
should not be allowed, so now the program checks this flag before
proceeding. You can still simulate defragmentation of unclean
filesystems.

The program now also marks filesystems as unclean when opening the disk,
and clean when closing it (again, unless simulating).