Added journal functions and unit test
The program is now capable of writing blocks to the journal (though it
doesn't yet). The new unit test checks this by writing a single block to
the journal without writing it to the actual data block, and then checks
that e2fsck properly replays it to the correct location.