Added journal testcase, fixed bugs found
There was a problem when writing a non-mmapped journal (the extent size
computed was off-by-one).

There were also bugs in journal_flush_upto: First it got confused by
DONE transactions at the start (assumed these still has to be flushed).
Second, the intended transaction can disappear while it's flushing the
preceding transactions (because it calls sync_disk), and it blew up when
it reached the end of the transaction list.

Finally, there was a stack overflow in journal_ensure_unprotected (I
figured there wouldn't be that many transactions protecting one block,
but in the new testcase there were).