Silence valgrind
We were writing a partially unitialized block to disk (specifically, the
extent descriptor block was only initialized as far as it was actually
used). Since this presents a possible security leak (and an annoying
valgrind warning), we can just use calloc to ensure the rest of the
block is all zeroes.